use int for precision parameter uniformly

+++ include/gradido_blockchain/GradidoUnit.h
@@ -22,7 +22,7 @@ public:
 	//! will be understood as gdd cent with 4 after comma, so gdd = gddCent / 10000
 	static GradidoUnit fromGradidoCent(int64_t gddCent) { return GradidoUnit(gddCent); }
 
-	GradidoUnit roundToPrecision(uint8_t precision = 4) const;
+	GradidoUnit roundToPrecision(int precision = 4) const;
 
 	//! \param precision expect value in the range [0;4]
 	std::string toString(int precision = 4) const;
